just a huge commit. add multi-tab support, re-organize repo, and add support for search pages
