+++ shared.js
@@ -1,14 +1,7 @@
-let _api = null;
-try {
-	_api = browser;
-	// manifest v2 has the action api stored in browserAction, so manually assign it to action
-	_api.action = browser.browserAction;
-}
-catch (ReferenceError) {
-	_api = chrome;
-}
-export const api = _api;
-export const logstr = "[Blue Blocker]";
+import { BlockCounter } from "./models/block_counter.js";
+import { BlockQueue } from "./models/block_queue.js";
+import { QueueConsumer } from "./models/queue_consumer.js";
+import { api, DefaultOptions, logstr, Headers, ReasonBlueVerified, ReasonNftAvatar, ReasonMap } from "./constants.js";
 
 var s = document.createElement("script");
 s.src = api.runtime.getURL("inject.js");
@@ -16,79 +9,12 @@ s.id = "injected-blue-block-xhr";
 s.type = "text/javascript";
 (document.head || document.documentElement).appendChild(s);
 
-export const DefaultOptions = {
-	// by default, spare as many people as possible
-	// let the user decide if they want to be stricter
-	blockFollowing: false,
-	blockFollowers: false,
-	skipVerified: true,
-	skipAffiliated: true,
-	skip1Mplus: true,
-	blockNftAvatars: false,
-	mute: false,
-	blockInterval: 10,
-};
+// Define constants that shouldn't be exported to the rest of the addon
+const queue = new BlockQueue(api.storage.local);
+const blockCounter = new BlockCounter(api.storage.local);
+const blockCache = new Set();
 
-// when parsing a timeline response body, these are the paths to navigate in the json to retrieve the "instructions" object
-// the key to this object is the capture group from the request regex in inject.js
-export const InstructionsPaths = {
-	HomeLatestTimeline: [
-		"data",
-		"home",
-		"home_timeline_urt",
-		"instructions",
-	],
-	HomeTimeline: [
-		"data",
-		"home",
-		"home_timeline_urt",
-		"instructions",
-	],
-	UserTweets: [
-		"data",
-		"user",
-		"result",
-		"timeline_v2",
-		"timeline",
-		"instructions",
-	],
-	TweetDetail: [
-		"data",
-		"threaded_conversation_with_injections_v2",
-		"instructions",
-	],
-};
-// this is the path to retrieve the user object from the individual tweet
-export const UserObjectPath = [
-	"tweet_results",
-	"result",
-	"tweet",
-	"core",
-	"user_results",
-	"result",
-];
-export const IgnoreTweetTypes = new Set([
-	"TimelineTimelineCursor",
-]);
-export const Headers = [
-	"authorization",
-	"x-twitter-active-user",
-	"x-twitter-auth-type",
-	"x-twitter-client-language",
-];
-
-// 64bit refid
-const MaxId = 0xffffffffffffffff;
-const RefId = () => Math.round(Math.random() * MaxId);
-
-export function commafy(x)
-{ // from https://stackoverflow.com/a/2901298
-	let parts = x.toString().split('.');
-	parts[0] = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-	return parts.join('.');
-}
-
-var options = { ...DefaultOptions };
+export var options = { ...DefaultOptions };
 export function SetOptions(items) {
 	options = items;
 }
@@ -96,138 +22,41 @@ export function SetOptions(items) {
 // retrieve settings immediately on startup
 api.storage.sync.get(DefaultOptions).then(SetOptions);
 
-const ReasonBlueVerified = 0;
-const ReasonNftAvatar = 1;
-
-const ReasonMap = {
-	[ReasonBlueVerified]: "Twitter Blue verified",
-	[ReasonNftAvatar]: "NFT avatar",
-};
-
-export class BlockQueue {
-	// queue must be defined with push and shift functions
-	constructor(storage) {
-		this.storage = storage;
-		this.queue = [];
-		this.timeout = null;
-	}
-	async sync() {
-		// sync simply adds the in-memory queue to the stored queue
-		const items = await this.storage.get({ BlockQueue: [] });
-		items.BlockQueue.push(...this.queue);
-		await this.storage.set(items);
-		this.queue.length = 0;
-		this.timeout = null;
-	}
-	async push(item) {
-		this.queue.push(item);
-		if (this.timeout) {
-			clearTimeout(this.timeout);
-		}
-		this.timeout = setTimeout(() => this.sync(), 100);
-	}
-	async shift() {
-		// shift halts any modifications to the local storage queue, removes an item, and saves it, and restarts sync
-		if (this.timeout) {
-			clearTimeout(this.timeout);
-		}
-		const items = await this.storage.get({ BlockQueue: [] });
-		const item = items.BlockQueue.shift();
-		if (item !== undefined) {
-			await this.storage.set(items);
-		}
-		this.timeout = setTimeout(() => this.sync(), 100);
-		return item;
-	}
-}
-
-export class BlockCounter {
-	// this class provides functionality to update and maintain a counter on badge text in an accurate way via async functions
-	constructor(storage) {
-		this.storage = storage;
-		this.value = 0;
-		this.timeout = null;
-
-		// we need to make sure the critical point is empty on launch. this has a very low chance of causing conflict between tabs, but
-		// prevents the possibility of a bunch of bugs caused by issues in retrieving the critical point. ideally we wouldn't have this
-		this.releaseCriticalPoint();
-	}
-	async getCriticalPoint() {
-		const key = "blockCounterCriticalPoint";
-		const refId = RefId();
-		let value = null;
-		do {
-			value = (await this.storage.get({ [key]: null }))[key];
-			if (!value) {
-				// try to access the critical point
-				await this.storage.set({ [key]: refId });
-				value = (await this.storage.get({ [key]: null }))[key];
-			}
-			else {
-				// sleep for a little bit to let the other tab(s) release the critical point
-				await new Promise(r => setTimeout(r, 50));
-			}
-		} while (value !== refId)
-	}
-	async releaseCriticalPoint() {
-		// this should only be called AFTER getCriticalPoint
-		const key = "blockCounterCriticalPoint";
-		await this.storage.set({ [key]: null });
-	}
-	async sync() {
-		await this.getCriticalPoint();
-		const items = await this.storage.get({ BlockCounter: 0 });
-		items.BlockCounter += this.value;
-		this.value = 0;
-		await this.storage.set(items);
-		this.releaseCriticalPoint();
-	}
-	async increment(value = 1) {
-		this.value += value;
-		if (this.timeout) {
-			clearTimeout(this.timeout);
-		}
-		this.timeout = setTimeout(() => this.sync(), 100);
-	}
-}
-
-const queue = new BlockQueue(api.storage.local);
-const blockCounter = new BlockCounter(api.storage.local);
-const BlockCache = new Set();
 let BlockTimeout = null;
 
 export function ClearCache() {
-	BlockCache.clear();
+	blockCache.clear();
 }
 
 function QueueBlockUser(user, user_id, headers, reason) {
-	if (BlockCache.has(user_id)) {
+	if (blockCache.has(user_id)) {
 		return;
 	}
-	BlockCache.add(user_id);
+	blockCache.add(user_id);
 	queue.push({user, user_id, headers, reason});
 	console.log(logstr, `queued ${user.legacy.name} (@${user.legacy.screen_name}) for a block due to ${ReasonMap[reason]}.`);
 
-	if (BlockTimeout === null) {
-		BlockTimeout = setTimeout(CheckBlockQueue, options.blockInterval * 1000);
-	}
+	consumer.start();
 }
 
 function CheckBlockQueue() {
 	queue.shift().then(item => {
 		if (item === undefined) {
-			clearTimeout(BlockTimeout);
-			BlockTimeout = null;
+			consumer.stop();
 			return;
 		}
 		api.storage.sync.get(DefaultOptions).then(items => {
 			SetOptions(items);
-			BlockTimeout = setTimeout(CheckBlockQueue, options.blockInterval * 1000);
 			const {user, user_id, headers, reason} = item;
 			BlockUser(user, user_id, headers, reason);
 		});
 	});
 }
+
+const consumer = new QueueConsumer(api.storage.local, CheckBlockQueue, async s => {
+	return (await s.get({ blockInterval: DefaultOptions.blockInterval })).blockInterval * 1000
+});
+consumer.start();
 
 const CsrfTokenRegex = /ct0=\s*(\w+);/;
 function BlockUser(user, user_id, headers, reason, attempt=1) {
@@ -239,10 +68,9 @@ function BlockUser(user, user_id, headers, reason, attempt=1) {
 	ajax.addEventListener('load', event => {
 		if (event.target.status === 403) {
 			// user has been logged out, we need to stop queue and re-add
-			clearTimeout(BlockTimeout);
-			BlockTimeout = null;
+			consumer.stop();
 			queue.push({user, user_id, headers, reason});
-			console.log(logstr, "user is logged out, queue consumer has been ceased.");
+			console.log(logstr, "user is logged out, queue consumer has been halted.");
 			return;
 		}
 		else if (event.target.status >= 300) {
@@ -308,13 +136,14 @@ export function BlockBlueVerified(user, headers) {
 		}
 		else if (
 			// group for skip-verified option
-			options.skipVerified && (user.legacy.verified || user.legacy.verified_type)
+			// TODO: look to see if there's some other way to check legacy verified
+			options.skipVerified && (user.legacy.verified)
 		) {
 			console.log(logstr, `did not block Twitter Blue verified user ${user.legacy.name} (@${user.legacy.screen_name}) because they are verified through other means.`);
 		}
 		else if (
 			// verified via an affiliated organisation instead of blue
-			options.skipAffiliated && user.affiliates_highlighted_label.label
+			options.skipAffiliated && (user?.affiliates_highlighted_label?.label || user.legacy.verified_type === "Business")
 		) {
 			console.log(logstr, `did not block Twitter Blue verified user ${user.legacy.name} (@${user.legacy.screen_name}) because they are verified through an affiliated organisation.`);
 		}
@@ -328,7 +157,7 @@ export function BlockBlueVerified(user, headers) {
 			QueueBlockUser(user, String(user.rest_id), headers, ReasonBlueVerified);
 		}
 	}
-	if (options.blockNftAvatars && user.has_nft_avatar) {
+	else if (options.blockNftAvatars && user.has_nft_avatar) {
 		if (
 			// group for block-following option
 			!options.blockFollowing && (user.legacy.following || user.super_following)
@@ -344,133 +173,5 @@ export function BlockBlueVerified(user, headers) {
 		else {
 			QueueBlockUser(user, String(user.rest_id), headers, ReasonNftAvatar);
 		}
-	}
-}
-
-function HandleTweetObject(obj, headers) {
-	let ptr = obj;
-	for (const key of UserObjectPath) {
-		if (ptr.hasOwnProperty(key)) {
-			ptr = ptr[key];
-		}
-	}
-	if (ptr.__typename !== "User") {
-		console.error(logstr, "could not parse tweet", obj);
-		return;
-	}
-	BlockBlueVerified(ptr, headers);
-}
-
-export function ParseTimelineTweet(tweet, headers) {
-	if(tweet.itemType=="TimelineTimelineCursor") {
-		return;
-	}
-	
-	// Handle retweets and quoted tweets (check the retweeted user, too)
-	if(tweet?.tweet_results?.result?.quoted_status_result) {
-		HandleTweetObject(tweet.tweet_results.result.quoted_status_result.result, headers);
-	} else if(tweet?.tweet_results?.result?.legacy?.retweeted_status_result) {
-		HandleTweetObject(tweet.tweet_results.result.legacy.retweeted_status_result.result, headers);
-	}
-	HandleTweetObject(tweet, headers);
-}
-
-export function HandleInstructionsResponse(e, body) {
-	// pull the "instructions" object from the tweet
-	let instructions = body;
-	
-	try {
-		for (const key of InstructionsPaths[e.detail.parsedUrl[1]]) {
-			instructions = instructions[key];
-		}
-	}
-	catch (e) {
-		console.error(logstr, "failed to parse response body for instructions object", e, body);
-		return;
-	}
-
-	// "instructions" should be an array, we need to iterate over it to find the "TimelineAddEntries" type
-	let tweets = undefined;
-	let isAddToModule = false;
-	for (const value of instructions) {
-		if (value.type === "TimelineAddEntries" || value.type === "TimelineAddToModule") {
-			tweets = value;
-			isAddToModule = value.type === "TimelineAddToModule";
-			break;
-		}
-	}
-	if (tweets === undefined) {
-		console.error(logstr, "response object does not contain an instruction to add entries", body);
-		return;
-	}
-
-	if (isAddToModule) {
-		// wrap AddToModule info so the handler can treat it the same (and unwrap it below)
-		tweets.entries = [{
-			content: {
-				entryType: "TimelineTimelineModule",
-				items: tweets.moduleItems
-			}
-		}];
-	}
-
-	// tweets object should now contain an array of all returned tweets
-	for (const tweet of tweets.entries) {
-		// parse each tweet for the user object
-		switch (tweet?.content?.entryType) {
-			case null:
-				console.error(logstr, "tweet structure does not match expectation", tweet);
-				break;
-
-			case "TimelineTimelineItem":
-				if (tweet.content.itemContent.itemType=="TimelineTweet") {
-					ParseTimelineTweet(tweet.content.itemContent, e.detail.request.headers);
-				}
-				break;
-			case "TimelineTimelineModule":
-				for (const innerTweet of tweet.content.items) {
-					ParseTimelineTweet(innerTweet.item.itemContent, e.detail.request.headers)
-				}
-				break;
-
-			default:
-				if (!IgnoreTweetTypes.has(tweet.content.entryType)) {
-					console.error(logstr, `unexpected tweet type found: ${tweet.content.entryType}`, tweet);
-				}
-		}
-	}
-
-	if (isAddToModule) {
-		tweets.moduleItems = tweets.entries[0]?.content?.items || [];
-		delete tweets.entries;
-	}
-}
-
-export function HandleHomeTimeline(e, body) {
-	// This API endpoint currently does not deliver information required for
-	// block filters (in particular, it's missing affiliates_highlighted_label).
-	// So if the user has set the "skip users verified by other means" options,
-	// this function must be skipped, however, it is still mostly covered by the
-	// instructions responses
-	if (options.skipAffiliated) return;
-
-	// so this url straight up gives us an array of users, so just use that lmao
-	for (const [user_id, user] of Object.entries(body.globalObjects.users)) {
-		// the user object is a bit different, so reshape it a little
-		BlockBlueVerified({
-			is_blue_verified: user.ext_is_blue_verified,
-			has_nft_avatar: user.ext_has_nft_avatar,
-			legacy: {
-				blocking: user.blocking,
-				followed_by: user.followed_by,
-				following: user.following,
-				name: user.name,
-				screen_name: user.screen_name,
-				verified: user.verified,
-				verified_type: user.ext_verified_type,
-			},
-			super_following: user.ext?.superFollowMetadata?.r?.ok?.superFollowing,
-			rest_id: user_id,
-		}, e.detail.request.headers)
 	}
 }
